+++ CHANGELOG.md
@@ -5,3 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+### Added
+- Test cases added for generation of GraphQL Schema from SHACL.
+
+### Removed
+- Trivial test case was removed, in favor of actual test cases.
+
+### Changed
+- The `lib` folder content was pulled up a directory.
+- The `model` package is now at the root level.